Applicate modifiche su autorizzazioni azioni su post e commenti

+++ faccia-libro/Backend/controllers/posts.js
@@ -50,8 +50,9 @@ export async function editPost(req, res, next) {
 
         //RIPARTI DA QUI
         const post = await Post.findById(id)
-        if(req.user._id !== post.author) {
-            
+        console.log(post)
+        if (!post.author.equals(req.user._id)) {
+            return res.status(403).json({ message: 'Non puoi modificare i post degli altri utenti' })
         }
 
         const editedPost = await Post.findByIdAndUpdate(id, {
@@ -72,13 +73,18 @@ export async function deletePost(req, res, next) {
     try {
         const { id } = req.params
 
-        const deletedUser = await Post.findByIdAndDelete(id)
+        const post = await Post.findById(id)
+        if (!post.author.equals(req.user._id)) {
+            return res.status(403).json({ message: 'Non puoi eliminare i post degli altri utenti' })
+        }
 
-        if (!deletedUser) {
+        const deletePost = await Post.findByIdAndDelete(id)
+
+        if (!deletePost) {
             return res.status(404).json({ message: 'Post non trovato' })
         }
 
-        res.status(200).json(deletedUser)
+        res.status(200).json(deletePost)
     } catch (err) {
         next(err)
     }
